refactor: extract seed-file writing into seedInterfaceFiles helper
